rename test-card to card
